Separate type-related pieces of numpy interop header into a separate header.
